Fix integer overflow in remote_bootstrap_client-test
Fixes the following Clang warning:

[823/936] Building CXX object src/kudu/tserver/CMakeFiles/remote_bootstrap_client-test.dir/remote_bootstrap_client-test.cc.o
src/kudu/tserver/remote_bootstrap_client-test.cc:149:29: warning: implicit conversion from 'unsigned long' to 'const int' changes value from 18446744073709551615 to -1 [-Wconstant-conversion]
  const int kDataTotalLen = std::numeric_limits<uint64_t>::max();
            ~~~~~~~~~~~~~   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
1 warning generated.

Reviewed-on: http://gerrit.cloudera.org:8080/1883
Tested-by: Todd Lipcon <todd@apache.org>

+++ src/kudu/tserver/remote_bootstrap_client-test.cc
@@ -146,7 +146,7 @@ TEST_F(RemoteBootstrapClientTest, TestVerifyData) {
   string bad = "This is a known bad! string";
   const int kGoodOffset = 0;
   const int kBadOffset = 1;
-  const int kDataTotalLen = std::numeric_limits<uint64_t>::max(); // Ignored.
+  const int64_t kDataTotalLen = std::numeric_limits<int64_t>::max(); // Ignored.
 
   // Create a known-good PB.
   DataChunkPB valid_chunk;
